refactor: 신고 생성 API 명세 수정 및 테스트 분리
