+++ backend/src/test/resources/features/report-member.feature
@@ -3,6 +3,6 @@ Feature: 유저를 신고한다
   Scenario: 정상적으로 다른 유저를 신고한다
     Given "jinwoo"의 깃허브 아이디로 회원가입을 한다.
     Given "noiman"의 깃허브 아이디로 회원가입을 한다.
-    Given 신고 제목-"노이만 신고", 내용-"코딩 너무 잘함", 문제발생일-"3"일 전으로 신고 정보를 작성한다.
-    When "jinwoo"가 제목이 "노이만 신고"인 신고 정보를 통해 "noiman"을 신고한다.
+    Given 피신고자-"noiman", 신고 제목-"노이만 신고", 내용-"코딩 너무 잘함", 문제발생일-"3"일 전으로 신고 정보를 작성한다.
+    When "jinwoo"가 제목이 "노이만 신고"인 신고 정보를 통해 신고 요청을 한다.
     Then 신고에 성공한다.
